backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.0.1
motor==3.3.1
pymongo==4.5.0
pydantic>=2.10.0
email-validator==2.1.1
python-jose[cryptography]==3.3.0
pyjwt==2.10.1
bcrypt==4.1.3
passlib==1.7.4
aiosmtplib==3.0.2
httpx>=0.28.1
openpyxl==3.1.2
python-multipart==0.0.9

# Development/Testing (optional, but good practice)
pytest==8.0.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
google-genai==1.68.0
google-generativeai==0.8.3
elevenlabs==2.39.0